requirements.txt
beautifulsoup4==4.12.3
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
psycopg2-binary==2.9.9
jinja2==3.1.4
python-multipart==0.0.12
httpx==0.27.2
anthropic==0.39.0
python-telegram-bot==21.7
google-api-python-client==2.149.0
google-auth-oauthlib==1.2.1
playwright==1.48.0
python-dotenv==1.0.1
pydantic-settings==2.6.0
eval_type_backport==0.2.0
pytest==8.3.3
pytest-asyncio==0.24.0
